# Fix misleading error display in /update-documentation command #80