--- a/.agent-os/specs/2025-08-27-fix-update-documentation-command-error-#80/spec.md
+++ b/.agent-os/specs/2025-08-27-fix-update-documentation-command-error-#80/spec.md
@@ -1,0 +1,106 @@
+# Spec Requirements Document
+
+> Spec: Fix Misleading Error Display in /update-documentation Command
+> Created: 2025-08-27
+> Status: Planning
+> Issue: #80
+
+## Overview
+
+The `/update-documentation` command currently displays misleading error messages when functioning correctly due to a mismatch between the script's intentional use of exit codes for signaling (exit code 2 = "documentation updates needed") and Claude Code's expectation that only exit code 0 indicates success. This creates a confusing user experience where successful operations appear to fail with "Error: Bash command failed for pattern" messages.
+
+This specification addresses the core issue by implementing a wrapper script pattern that translates semantic exit codes into user-friendly status messages while maintaining compatibility with both Claude Code's command system and existing CI/CD workflows.
+
+## User Stories
+
+**As a developer using Agent OS,**
+I want the `/update-documentation` command to display clear, accurate status messages
+So that I can understand whether the command succeeded and what action (if any) is needed.
+
+**As a CI/CD system,**
+I want to continue receiving the original semantic exit codes from update-documentation.sh
+So that automated workflows can make appropriate decisions based on documentation status.
+
+**As a future Agent OS maintainer,**
+I want a documented pattern for handling semantic exit codes in Claude Code commands
+So that similar issues can be avoided and resolved consistently.
+
+## Spec Scope
+
+### In Scope
+- Create wrapper script for `/update-documentation` command that handles exit code translation
+- Implement clear status messaging for different documentation states
+- Preserve original script behavior for direct CLI and CI/CD usage
+- Document the exit code wrapper pattern for future reference
+- Ensure all existing functionality is maintained
+
+### Out of Scope
+- Modifying the core update-documentation.sh logic or behavior
+- Changing CI/CD workflows or automation that depends on current exit codes
+- Implementing similar fixes for other commands (this is a pattern proof-of-concept)
+- UI/visual improvements beyond message clarity
+
+## Success Criteria
+
+1. **No False Errors**: The `/update-documentation` command shows no error messages when working correctly
+2. **Clear Status Messages**: Users receive explicit feedback about documentation status:
+   - "Documentation is up-to-date" when no changes needed
+   - "Documentation updates recommended" when changes detected
+3. **Backward Compatibility**: Original update-documentation.sh maintains current behavior for direct usage
+4. **CI/CD Compatibility**: Automated systems continue receiving semantic exit codes as expected  
+5. **Pattern Documentation**: Exit code wrapper pattern is documented for future use
+6. **Complete Functionality**: All existing command features and options work unchanged
+
+## Non-Goals
+
+- Redesigning the update-documentation workflow or logic
+- Creating a universal exit code handling system for all Agent OS commands
+- Modifying Claude Code's command parsing or error handling behavior
+- Performance optimization of the documentation checking process
+
+## Assumptions
+
+- Claude Code will continue using the `!` syntax for command execution with current error handling
+- Existing CI/CD workflows depend on the current exit code behavior and should not be disrupted  
+- The context-aware-wrapper.sh pattern provides a suitable foundation for this implementation
+- Users primarily interact with update-documentation through Claude Code rather than direct CLI
+- The issue is specific to Claude Code interaction and doesn't affect other AI tools
+
+## Dependencies
+
+- **Technical**: Access to existing context-aware-wrapper.sh implementation for pattern reference
+- **Workflow**: Must follow Agent OS development workflow (source → install → deploy)  
+- **Documentation**: Update to Claude Code command documentation may be needed
+- **Testing**: Verification with both Claude Code and direct CLI usage
+
+## Risks
+
+### Technical Risks
+- **Wrapper Complexity**: Additional layer could introduce unexpected behavior or edge cases
+- **Performance Impact**: Extra wrapper execution could slow command response time
+- **Maintenance Burden**: Additional script to maintain and test
+
+### Process Risks  
+- **Breaking Changes**: Risk of inadvertently changing behavior that other systems depend on
+- **Incomplete Coverage**: Missing edge cases where the wrapper doesn't handle exit codes properly
+- **Documentation Drift**: Pattern documentation becoming outdated as implementation evolves
+
+### Mitigation Strategies
+- Thorough testing with both interactive and automated usage patterns
+- Clear separation between wrapper and original script responsibilities
+- Comprehensive documentation with examples and troubleshooting guidance
+
+## Expected Deliverable
+
+A complete solution that eliminates misleading error messages in the `/update-documentation` command while preserving all existing functionality, consisting of:
+
+1. **Wrapper Script**: New wrapper following context-aware-wrapper.sh pattern
+2. **Command Integration**: Updated Claude Code command configuration  
+3. **Documentation**: Pattern documentation for future reference
+4. **Testing**: Verification that all functionality works in both Claude Code and CI/CD contexts
+
+## Spec Documentation
+
+- Tasks: @.agent-os/specs/2025-08-27-fix-update-documentation-command-error-#80/tasks.md
+- Technical Specification: @.agent-os/specs/2025-08-27-fix-update-documentation-command-error-#80/sub-specs/technical-spec.md
+- Implementation Guide: @.agent-os/specs/2025-08-27-fix-update-documentation-command-error-#80/sub-specs/implementation-guide.md