--- a/.agent-os/specs/2025-08-27-fix-update-documentation-command-error-#80/sub-specs/technical-spec.md
+++ b/.agent-os/specs/2025-08-27-fix-update-documentation-command-error-#80/sub-specs/technical-spec.md
@@ -1,0 +1,287 @@
+# Technical Specification
+
+This is the technical specification for the spec detailed in @.agent-os/specs/2025-08-27-fix-update-documentation-command-error-#80/spec.md
+
+> Created: 2025-08-27
+> Version: 1.0.0
+
+## Problem Analysis
+
+### Root Cause
+
+The `/update-documentation` command displays misleading error messages due to a semantic mismatch between:
+
+1. **Script Behavior**: `update-documentation.sh` uses exit code 2 to signal "documentation updates needed" (not an error)
+2. **Claude Code Expectation**: Claude Code interprets any non-zero exit code as command failure
+3. **User Experience**: Users see "Error: Bash command failed" when the command is working correctly
+
+### Current Exit Code Semantics
+
+From analysis of `update-documentation.sh`:
+- `exit 0`: No documentation updates needed OR diff-only mode OR normal success
+- `exit 2`: Documentation updates required (lines 312, 323) - **semantic success, not failure**
+
+### Current Command Flow
+
+```
+Claude Code → commands/update-documentation.md → !`~/.agent-os/scripts/update-documentation.sh $ARGUMENTS` → exit 2 → "Error: Bash command failed"
+```
+
+## Technical Approach Options
+
+### Option 1: Exit Code Wrapper Script (Recommended)
+
+**Approach**: Create a wrapper script that translates semantic exit codes into user-friendly messages and always returns exit 0 to Claude Code.
+
+**Benefits**:
+- Preserves existing script behavior for CI/CD and direct usage
+- Eliminates misleading error messages in Claude Code
+- Follows existing Agent OS pattern (context-aware-wrapper.sh)
+- Minimal complexity and maintenance burden
+- Clear separation of concerns
+
+**Architecture**:
+```
+Claude Code → wrapper → update-documentation.sh → wrapper interprets exit code → friendly message + exit 0
+```
+
+### Option 2: Command Configuration Modification
+
+**Approach**: Modify the Claude Code command definition to handle non-zero exit codes gracefully.
+
+**Benefits**:
+- No additional scripts needed
+- Direct integration with Claude Code
+
+**Drawbacks**:
+- Limited control over Claude Code's error handling behavior
+- May not fully eliminate error message display
+- Less portable to other AI tools
+
+### Option 3: Core Script Modification
+
+**Approach**: Modify `update-documentation.sh` to use exit 0 for all success cases and reserve non-zero codes for true errors.
+
+**Benefits**:
+- Simplest solution
+- No wrapper complexity
+
+**Drawbacks**:
+- **Breaking change** for CI/CD workflows that depend on exit 2
+- Violates backward compatibility requirement
+- Could break automated systems
+
+## Recommended Solution: Exit Code Wrapper
+
+### Implementation Details
+
+#### Wrapper Script: `update-documentation-wrapper.sh`
+
+```bash
+#!/usr/bin/env bash
+# Wrapper for update-documentation.sh that provides Claude Code-friendly output
+# Location: ~/.agent-os/scripts/update-documentation-wrapper.sh
+
+set -euo pipefail
+
+# Execute the original script and capture exit code
+if output=$("$HOME/.agent-os/scripts/update-documentation.sh" "$@" 2>&1); then
+    exit_code=$?
+else
+    exit_code=$?
+fi
+
+# Always display the output
+echo "$output"
+
+# Translate exit codes to user-friendly messages
+case $exit_code in
+    0)
+        echo ""
+        echo "✅ Documentation is up-to-date"
+        ;;
+    2)
+        echo ""
+        echo "📝 Documentation updates recommended (see suggestions above)"
+        ;;
+    *)
+        echo ""
+        echo "❌ Documentation check failed with exit code $exit_code"
+        # For true errors, preserve the exit code
+        exit $exit_code
+        ;;
+esac
+
+# Always exit 0 for semantic success cases (0 and 2)
+exit 0
+```
+
+#### Command Configuration Update
+
+Modify `commands/update-documentation.md` line 18:
+```markdown
+!`~/.agent-os/scripts/update-documentation-wrapper.sh $ARGUMENTS`
+```
+
+#### Installation Integration
+
+Update `setup.sh` or installation scripts to deploy the wrapper:
+```bash
+# Copy wrapper script
+cp scripts/update-documentation-wrapper.sh "$HOME/.agent-os/scripts/"
+chmod +x "$HOME/.agent-os/scripts/update-documentation-wrapper.sh"
+```
+
+### Architecture Considerations
+
+#### Script Relationship
+```
+Source Files:
+  scripts/update-documentation.sh         (core logic, unchanged)
+  scripts/update-documentation-wrapper.sh (new wrapper)
+  commands/update-documentation.md        (updated command reference)
+
+Installation:
+  ~/.agent-os/scripts/update-documentation.sh         (deployed core)
+  ~/.agent-os/scripts/update-documentation-wrapper.sh (deployed wrapper)
+  ~/.claude/commands/update-documentation.md          (updated command)
+
+Execution Flow:
+  Claude Code → wrapper → core script → wrapper translates → friendly output
+```
+
+#### Backward Compatibility Strategy
+
+**Direct CLI Usage**:
+- `update-documentation.sh` remains unchanged
+- Users can still call original script directly
+- CI/CD workflows continue to work with exit 2 semantics
+
+**Claude Code Usage**:
+- Uses wrapper for friendly output
+- No error messages for successful operations
+- Clear status messaging for all scenarios
+
+#### Error Handling Strategy
+
+**Wrapper Responsibilities**:
+- Capture and display all original script output
+- Translate exit codes 0 and 2 to friendly messages with exit 0
+- Preserve non-zero exit codes for true errors (1, 3+)
+- Add visual indicators (✅, 📝, ❌) for clarity
+
+**Original Script Unchanged**:
+- Maintains all existing logic and exit code semantics
+- No risk of breaking existing CI/CD integrations
+- Preserves all command-line arguments and functionality
+
+#### Performance Considerations
+
+**Overhead Analysis**:
+- Wrapper adds minimal overhead (single script execution)
+- No additional external dependencies
+- Uses standard bash features only
+- Expected performance impact: < 10ms additional latency
+
+**Optimization Strategies**:
+- Direct exec to original script where possible
+- Minimal string processing in wrapper
+- No additional file I/O beyond original script
+
+## Dependencies
+
+### Technical Dependencies
+- **Bash**: Standard bash shell features (set -euo pipefail, case statements, exit code capture)
+- **File System**: Write access to ~/.agent-os/scripts/ directory
+- **Original Script**: Existing update-documentation.sh must remain functional
+
+### Integration Dependencies
+- **Setup Scripts**: Integration with setup.sh or setup-claude-code.sh for deployment
+- **Command System**: Claude Code command configuration system
+- **Version Control**: Integration with Agent OS source → install → deploy workflow
+
+### External Dependencies
+- **Git**: Required by original update-documentation.sh (unchanged)
+- **GitHub CLI**: Optional dependency for enhanced features (unchanged)
+- **jq**: Optional for JSON parsing (unchanged)
+
+## Implementation Strategy
+
+### Phase 1: Core Wrapper Development
+1. Create wrapper script with exit code translation
+2. Test wrapper with all update-documentation.sh modes
+3. Verify output formatting and messaging
+4. Test error handling for true failure cases
+
+### Phase 2: Integration
+1. Update command configuration to use wrapper
+2. Integrate wrapper deployment into setup scripts
+3. Test end-to-end flow in Claude Code
+4. Verify backward compatibility for direct CLI usage
+
+### Phase 3: Documentation and Testing
+1. Document the exit code wrapper pattern
+2. Create test cases for wrapper functionality
+3. Update command documentation with new behavior
+4. Add wrapper to Agent OS health checks
+
+## Risk Mitigation
+
+### Technical Risks
+
+**Risk**: Wrapper introduces unexpected behavior
+- **Mitigation**: Comprehensive testing with all command modes and arguments
+- **Detection**: Unit tests for wrapper script
+- **Recovery**: Ability to revert command to direct script execution
+
+**Risk**: Output formatting changes break user workflows
+- **Mitigation**: Preserve all original output, only add friendly status messages
+- **Detection**: Before/after output comparison testing
+- **Recovery**: Wrapper script can be easily modified or bypassed
+
+**Risk**: Performance degradation
+- **Mitigation**: Minimal wrapper implementation with direct execution
+- **Detection**: Performance benchmarking before/after implementation
+- **Recovery**: Direct script execution remains available
+
+### Process Risks
+
+**Risk**: Breaking CI/CD workflows
+- **Mitigation**: Original script remains unchanged and available for direct use
+- **Detection**: CI/CD workflows should continue using original script path
+- **Recovery**: Original script functionality is completely preserved
+
+**Risk**: Installation complexity increases
+- **Mitigation**: Simple file copy with permissions, following existing patterns
+- **Detection**: Test installation process in clean environments
+- **Recovery**: Manual installation of wrapper script is straightforward
+
+## Success Metrics
+
+### Functional Metrics
+- No "Error: Bash command failed" messages when documentation check succeeds
+- Clear status messages for all operation outcomes
+- All existing command functionality preserved
+- Original script behavior unchanged for direct usage
+
+### User Experience Metrics
+- Reduced user confusion about command success/failure
+- Clear actionability of command output
+- Consistent behavior across different usage contexts
+
+### Technical Metrics
+- Performance impact < 10ms additional latency
+- No additional external dependencies introduced
+- 100% backward compatibility maintained
+- Zero breaking changes to existing workflows
+
+## Future Considerations
+
+### Pattern Reusability
+This exit code wrapper pattern could be applied to other Agent OS commands that use semantic exit codes, providing a foundation for consistent Claude Code integration.
+
+### Enhanced Messaging
+Future versions could include more sophisticated status messaging, progress indicators, or integration with Agent OS notification systems.
+
+### Multi-Tool Compatibility
+The wrapper pattern could be extended to provide consistent behavior across different AI coding tools beyond Claude Code.