… .agent-os/specs/2025-08-27-fix-update-documentation-command-error-#80/sub-specs/tests.md
… .agent-os/specs/2025-08-27-fix-update-documentation-command-error-#80/tasks.md
… PROJECT_INDEX.json
… docs/EXIT_CODE_WRAPPER_PATTERN.md